[REF] Delay start of the SendsIllFormattedString
Axel stated:
"What also did confuse me is, that this is the first test that runs - and it causes a
crash. When checking various logs investigating something completely different, that
made me confused if everything is right, because it looked like the test system is not
running at all. I'd recommend to change the test order also and start with some good case
tests first. Just to ensure things are basically working."

To improve this, now the sendsIllFormattedString will be delayed waiting for the ready event
from the app_lvl_max.

Fixes:
SEOS-1128